requirements.txt
streamlit
huggingface_hub
groq
python-dotenv
